requirements.txt
# GhostBit - Multi-Modal Steganography Framework
# Python 3.11+

# Web Framework
streamlit>=1.28.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Audio Processing
soundfile>=0.12.0

# Video Processing (bundled FFmpeg)
imageio>=2.30.0
imageio-ffmpeg>=0.4.9
av>=10.0.0

# Cryptography
cryptography>=41.0.0

# Metrics
scikit-image>=0.21.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Documentation
python-docx>=0.8.11

# Utilities
tqdm>=4.66.0
